+++ controllers/template.js
@@ -1,5 +1,7 @@
 /* eslint-disable no-unused-vars */
 export default class Controller {
+	initListeners () {}
+	
 	async send (message, options = {}) {}
 
 	destroy () {}